[#54] - Screen file structure & invite members screen
