+++ js/components/Root.tsx
@@ -10,8 +10,9 @@ import { connect, ConnectedProps } from 'react-redux'
 import { userSlice } from '../reducers/user'
 import { RootState } from '../redux/store'
 import { RootStackParamsT } from '../types/RootStackParams'
-import AuthenticationScreen from './Authentication/AuthenticationScreen'
-import CreatePlanScreen from './CreatePlan/CreatePlanScreen'
+import AuthenticationScreen from './Screens/Authentication/AuthenticationScreen'
+import CreatePlanScreen from './Screens/Modals/CreatePlan/CreatePlanScreen'
+import InvitePlanMembersScreen from './Screens/Modals/InvitePlanMembers/InviteMembersScreen'
 import SplashScreen from './SplashScreen'
 import TabNavigator from './TabNavigator'
 import ThemeProvider from './ThemeProvider'
@@ -75,6 +76,10 @@ const Root = (props: ReduxPropsT) => {
       <NavigationContainer>
         <RootStack.Navigator initialRouteName="tab" screenOptions={rootOptions}>
           <RootStack.Screen name="createPlan" component={CreatePlanScreen} />
+          <RootStack.Screen
+            name="inviteGroupPlanMembers"
+            component={InvitePlanMembersScreen}
+          />
           <RootStack.Screen name="tab" component={TabNavigator} />
         </RootStack.Navigator>
       </NavigationContainer>
