+++ js/components/Screens/DashboardTabNavigator.tsx
@@ -1,7 +1,7 @@
 import { createStackNavigator } from '@react-navigation/stack'
 import React, { memo, useMemo } from 'react'
-import { useAppTheme } from '../hooks/useAppTheme'
-import { DashboardTabNavigatorPropsT } from '../types/DashboardTabNavigatorProps'
+import { useAppTheme } from '../../hooks/useAppTheme'
+import { DashboardTabNavigatorPropsT } from '../../types/DashboardTabNavigatorProps'
 import UserScreen from './DashboardScreen'
 
 const Stack = createStackNavigator<DashboardTabNavigatorPropsT>()
