+++ js/components/Screens/Modals/CreatePlan/CreatePlanScreen.tsx
@@ -9,19 +9,19 @@ import { StackScreenProps } from '@react-navigation/stack'
 import React, { memo, useCallback, useMemo, useState } from 'react'
 import { View } from 'react-native'
 import { connect, ConnectedProps } from 'react-redux'
-import { planColors } from '../../constants/planColors'
-import { planIcons } from '../../constants/planIcons'
-import { useAppTheme } from '../../hooks/useAppTheme'
-import { plansSlice } from '../../reducers/plans'
-import { RootState } from '../../redux/store'
-import { ModalScreensParamsT } from '../../types/ModalScreensParams'
-import { assertNever } from '../../utils/assertNever'
-import { formatDate } from '../../utils/date'
-import { createPlanLoader } from '../../utils/loaders'
-import Button from '../Button'
-import ScrollContainer from '../ScrollContainer'
-import TextInput from '../TextInput'
-import VerticalSeparator from '../VerticalSepartor'
+import { planColors } from '../../../../constants/planColors'
+import { planIcons } from '../../../../constants/planIcons'
+import { useAppTheme } from '../../../../hooks/useAppTheme'
+import { plansSlice } from '../../../../reducers/plans'
+import { RootState } from '../../../../redux/store'
+import { ModalScreensParamsT } from '../../../../types/ModalScreensParams'
+import { assertNever } from '../../../../utils/assertNever'
+import { formatDate } from '../../../../utils/date'
+import { createPlanLoader } from '../../../../utils/loaders'
+import Button from '../../../Button'
+import ScrollContainer from '../../../ScrollContainer'
+import TextInput from '../../../TextInput'
+import VerticalSeparator from '../../../VerticalSepartor'
 import ColorSelector from './ColorSelector'
 import IconSelector from './IconSelector'
