+++ js/components/Screens/Plans/PlansScreen.tsx
@@ -8,15 +8,15 @@ import React, {
 } from 'react'
 import { FlatList, ListRenderItemInfo } from 'react-native'
 import { connect, ConnectedProps } from 'react-redux'
-import { useAppTheme } from '../../hooks/useAppTheme'
-import { plansSlice } from '../../reducers/plans'
-import { RootState } from '../../redux/store'
-import { ModalScreensParamsT } from '../../types/ModalScreensParams'
-import { PlanT } from '../../types/Plan'
-import { PlansTabNavigatorPropsT } from '../../types/PlansTabNavigatorProps'
-import { plansLoader } from '../../utils/loaders'
-import IconButton from '../IconButton'
-import ScreenWithLoader from '../ScreenWithLoader'
+import { useAppTheme } from '../../../hooks/useAppTheme'
+import { plansSlice } from '../../../reducers/plans'
+import { RootState } from '../../../redux/store'
+import { ModalScreensParamsT } from '../../../types/ModalScreensParams'
+import { PlanT } from '../../../types/Plan'
+import { PlansTabNavigatorPropsT } from '../../../types/PlansTabNavigatorProps'
+import { plansLoader } from '../../../utils/loaders'
+import IconButton from '../../IconButton'
+import ScreenWithLoader from '../../ScreenWithLoader'
 import PlanItem from './PlanItem'
 
 type NavigationPropsT = StackScreenProps<
