+++ js/types/ModalScreensParams.ts
@@ -1,3 +1,4 @@
 export type ModalScreensParamsT = {
   createPlan: undefined
+  inviteGroupPlanMembers: undefined
 }
